Placeholder to Production: At least for the data handling
POJO Updates
Ingredient.java: Changed int to double based on the data type (floating point).
Recipe.java: Changed Parcel calls to utilize the typed input and output functions.

POJO Handler
RecipeContent.java: Added a new 'global' glass to store the Recipe content, to be accessed throughout the application where needed.

RecipeDetailActivity.java
*Removed the placeholder calls and swapped them to utilize the new parcel object.

RecipeDetailFragment.java
*Stripped out the dummycontent in favor of recipe objects.
*Added additional null/sanity checks.

RecipeListActivity.java
*Added member variables for recycler, adapter, and observer.
*Added observer calls to track the recipe/changes, and populate if it changes (or changes after the UI has finished and then the network call finishes).
*Updated dummy calls to data calls.
*Added additional null/sanity checks.

+++ app/src/main/java/com/weeturretstudio/warbeleth/android/bakingapp/model/Ingredient.java
@@ -35,7 +35,7 @@ public class Ingredient implements Parcelable
             ;
 
     protected Ingredient(Parcel in) {
-        this.quantity = ((int) in.readValue((int.class.getClassLoader())));
+        this.quantity = ((double) in.readValue((double.class.getClassLoader())));
         this.measure = ((String) in.readValue((String.class.getClassLoader())));
         this.ingredient = ((String) in.readValue((String.class.getClassLoader())));
     }
